Change the way to find new items inside a feed
Before, we only check that we have the last item, if yes, we don't fetch items. Otherwise, we fetch items until we reach the cached one.

Now, we compare cached links and item links to be sure to fetch *all* new item.

It's a bit longer, but much more "sure".

+ rename feed `$id` parameter to `$feedId`