Rewrote the way to override site config when testing
Look likes that feature doesn't really works 🤔
At least, if the given domain already has a site config defined in graby-site-config, it was impossible to add more config to it.

Now, config which is tested is merged in the existing file (or the file is created) and after the test, the file is reverted (or deleted if it doesn't exist first).